algo/src/main/kotlin/com/hartwig/actin/algo/evaluation/vitalfunction/HasSufficientBodyWeight.kt
package com.hartwig.actin.algo.evaluation.vitalfunction

import com.hartwig.actin.PatientRecord
import com.hartwig.actin.algo.datamodel.Evaluation
import com.hartwig.actin.algo.evaluation.EvaluationFunction

class HasSufficientBodyWeight(val referenceWeight: Double) : EvaluationFunction {

    override fun evaluate(record: PatientRecord): Evaluation {
        return BodyWeightFunctions.evaluatePatientBodyWeightAgainstReference(
            record, this.referenceWeight, referenceIsMinimum = true
        )
    }
}